requirements.txt
python~=3.9.6
pandas~=1.4.2
pydicom~=2.3.0
numpy~=1.19.5
pillow~=9.0.1
keras~=2.9.0
tensorflow~=2.4.1
tensorflow-gpu~=2.9.1
opencv-python~=4.6.0.66
scipy~=1.7.3
matplotlib~=3.5.2

sklearn~=0.0
scikit-learn~=1.1.1
dataclasses~=0.8
nibabel~=4.0.1
tqdm~=4.64.0